requirements.txt
click==6.6
coverage==4.0.3
python-dotenv==0.4.0
oauthlib==1.0.3
pluggy==0.3.1
py==1.4.31
pytest==2.9.1
python-dateutil==2.5.3
python-env==1.0.0
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
tox==2.3.1
tweepy==3.5.0
-e git+git@github.com:EastAgile/twitterbot.git#egg=responsebot
virtualenv==15.0.1
